% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_in.R
\name{reading_in}
\alias{reading_in}
\title{Function for parsing the text output of FastQC}
\usage{
reading_in(file, test = "Per base sequence quality", sample_name)
}
\arguments{
\item{file}{string that specifies the path to an individual FastQC result file
(tyically named "fastqc_data.txt"}

\item{test}{Indicate which test results should be extracted. Default:
"Per base sequence quality". Other options are, for example, "Per tile sequence quality",
"Per sequence quality score" etc.}
}
\value{
data.frame with the values of a single FastQC test result.
}
\description{
This functions extracts the values for a specific test run by FastQC on a
single fastq file.
}
\examples{
\dontrun{
res <- reading_in(file = "acinar-3_S9_L001_R1_001_fastqc/fastqc_data.txt")
}
}
